+++ frb_example/dart_minimal/rust/src/api/minimal.rs
@@ -1,7 +1,4 @@
-use crate::frb_generated::{StreamSink, FLUTTER_RUST_BRIDGE_HANDLER};
-use anyhow::anyhow;
-use flutter_rust_bridge::for_generated::BaseThreadPool;
-use flutter_rust_bridge::{frb, transfer};
+use flutter_rust_bridge::frb;
 
 #[frb(init)]
 pub fn init_app() {
@@ -10,12 +7,4 @@ pub fn init_app() {
 
 pub fn minimal_adder(a: i32, b: i32) -> i32 {
     a + b
-}
-
-pub fn func_stream_add_value_and_error_twin_normal(sink: StreamSink<i32>) {
-    (FLUTTER_RUST_BRIDGE_HANDLER.thread_pool()).execute(transfer!(|| {
-        sink.add(100).unwrap();
-        sink.add(200).unwrap();
-        sink.add_error(anyhow!("deliberate error")).unwrap();
-    }));
 }
