+++ frb_example/dart_minimal/test/minimal_test.dart
@@ -16,20 +16,4 @@ Future<void> main() async {
     print('Action: Call rust (after)');
   });
   print('Action: Configure tests (end)');
-
-  test('func_stream_add_value_and_error_twin_normal', () async {
-    final stream = await funcStreamAddValueAndErrorTwinNormal();
-    final events = <String>[];
-    final onDone = Completer<void>();
-    stream.listen(
-      (e) => events.add('data $e'),
-      onError: (e, s) {
-        print('onError $e $s');
-        events.add('error $e');
-      },
-      onDone: () => onDone.complete(),
-    );
-    await onDone.future;
-    expect(events, ['data 100', 'data 200', contains('deliberate error')]);
-  });
 }
