+++ frb_example/pure_dart/rust/src/api/stream.rs
@@ -76,3 +76,11 @@ pub struct MyStructContainingStreamSinkTwinNormal {
 pub fn stream_sink_inside_struct_twin_normal(arg: MyStructContainingStreamSinkTwinNormal) {
     arg.b.add(arg.a).unwrap();
 }
+
+pub fn func_stream_add_value_and_error_twin_normal(sink: StreamSink<i32>) {
+    (FLUTTER_RUST_BRIDGE_HANDLER.thread_pool()).execute(transfer!(|| {
+        sink.add(100).unwrap();
+        sink.add(200).unwrap();
+        sink.add_error(anyhow!("deliberate error")).unwrap();
+    }));
+}
