+++ frb_example/pure_dart/test/api/stream_test.dart
@@ -89,4 +89,20 @@ Future<void> main({bool skipRustLibInit = false}) async {
     await streamSinkInsideStructTwinNormal(arg: arg);
     expect(await arg.b.stream.toList(), [1000]);
   });
+
+  test('func_stream_add_value_and_error_twin_normal', () async {
+    final stream = await funcStreamAddValueAndErrorTwinNormal();
+    final events = <String>[];
+    final onDone = Completer<void>();
+    stream.listen(
+      (e) => events.add('data $e'),
+      onError: (e, s) {
+        print('onError $e $s');
+        events.add('error $e');
+      },
+      onDone: () => onDone.complete(),
+    );
+    await onDone.future;
+    expect(events, ['data 100', 'data 200', contains('deliberate error')]);
+  });
 }
